Support URLs with ports
Match patterns must not include ports, else they won't work, but the URLs are
currently passed to the listener requestFilter unmodified.  Strip out the ports
beforehand instead, and then if we happen to get a listener match for the same
URL on a different port just pass it through rather than using the replacement
from another port.

Fixes #4